+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 4.0.0
+
+### Added
+- Added support for .NET 6.0.
+
+### Removed
+- Removed support for .NET Core 2.1 and 3.0.
+
+### Fixed
+- Fixed vulnerability "Cookie parsing failure" related to Microsoft.AspNetCore.Http.
+
 ## 3.0.0
 
 ### Added